fix: move entities to main and raw
