+++ frontend/src/features/auth/entities/index.ts
@@ -1,2 +1,1 @@
-export * from "./credential";
-export * from "./token";
+export * from "./main";
