+++ frontend/src/features/auth/useCases/loginUseCase.ts
@@ -1,5 +1,4 @@
-import { Credential } from "../entities/credential";
-import { Token } from "../entities/token";
+import { Credential, Token } from "../entities";
 
 export interface LoginService {
   loginWithCredential: (credential: Credential) => Promise<Token>;
